考虑辅导员查询方式，考虑教职工和院系的关系等。

+++ src/main/java/com/yunhuakeji/attendance/enums/RoleType.java
@@ -1,20 +1,20 @@
 package com.yunhuakeji.attendance.enums;
 
 public enum RoleType {
-    SecondaryCollegeAdmin(1, "二级院校管理员"),
-    DormitoryAdmin(2, "宿舍管理员"),
-    StudentsAffairsAdmin(3, "学生处管理员");
+    SecondaryCollegeAdmin((byte)1, "二级院校管理员"),
+    DormitoryAdmin((byte)2, "宿舍管理员"),
+    StudentsAffairsAdmin((byte)3, "学生处管理员");
 
 
-    private int type;
+    private byte type;
     private String msg;
 
-    RoleType(int type, String msg) {
+    RoleType(byte type, String msg) {
         this.type = type;
         this.msg = msg;
     }
 
-    public static RoleType get(int val) {
+    public static RoleType get(byte val) {
         for (RoleType c : RoleType.values()) {
             if (c.getType() == val) {
                 return c;
@@ -23,7 +23,7 @@ public enum RoleType {
         return null;
     }
 
-    public int getType() {
+    public byte getType() {
         return type;
     }
